remove load_dataframe method from base_table
