+++ src/teehr/evaluation/tables/location_attribute_table.py
@@ -13,6 +13,8 @@ import logging
 from teehr.utils.utils import to_path_or_s3path, remove_dir_if_exists
 from teehr.loading.utils import add_or_replace_sdf_column_prefix
 from teehr.models.table_enums import TableWriteEnum
+import pyspark.sql as ps
+import pandas as pd
 
 
 logger = logging.getLogger(__name__)
@@ -241,4 +243,59 @@ class LocationAttributeTable(BaseTable):
             drop_duplicates=drop_duplicates,
             **kwargs
         )
+        self._load_table()
+
+    def load_dataframe(
+        self,
+        df: Union[pd.DataFrame, ps.DataFrame],
+        field_mapping: dict = None,
+        constant_field_values: dict = None,
+        location_id_prefix: str = None,
+        write_mode: TableWriteEnum = "append",
+        persist_dataframe: bool = False,
+        drop_duplicates: bool = True,
+    ):
+        """Import data from an in-memory dataframe.
+
+        Parameters
+        ----------
+        df : Union[pd.DataFrame, ps.DataFrame]
+            DataFrame to load into the table.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        constant_field_values : dict, optional
+            A dictionary mapping field names to constant values.
+            Format: {field_name: value}.
+        location_id_prefix : str, optional
+            The prefix to add to location IDs.
+            Used to ensure unique location IDs across configurations.
+            Note, the methods for fetching USGS and NWM data automatically
+            prefix location IDs with "usgs" or the nwm version
+            ("nwm12, "nwm21", "nwm22", or "nwm30"), respectively.
+        write_mode : TableWriteEnum, optional (default: "append")
+            The write mode for the table.
+            Options are "append", "upsert", and "overwrite".
+            If "append", the table will be appended with new data that does
+            already exist.
+            If "upsert", existing data will be replaced and new data that
+            does not exist will be appended.
+            If "overwrite", existing partitions receiving new data are overwritten.
+        persist_dataframe : bool, optional (default: False)
+            Whether to repartition and persist the pyspark dataframe after
+            reading from the cache. This can improve performance when loading
+            a large number of files from the cache.
+        drop_duplicates : bool, optional (default: True)
+            Whether to drop duplicates from the dataframe.
+        """
+        self._load_dataframe(
+            df=df,
+            field_mapping=field_mapping,
+            constant_field_values=constant_field_values,
+            location_id_prefix=location_id_prefix,
+            write_mode=write_mode,
+            persist_dataframe=persist_dataframe,
+            drop_duplicates=drop_duplicates
+        )
+
         self._load_table()
